// src/types.ts
export type BookAbbr = 
   "gn"
 | "ex"
 | "lv"
 | "nm"
 | "dt"
 | "js"
 | "jz"
 | "rt"
 | "1sm"
 | "2sm"
 | "1rs"
 | "2rs"
 | "1cr"
 | "2cr"
 | "ed"
 | "ne"
 | "et"
 | "jó"
 | "sl"
 | "pv"
 | "ec"
 | "ct"
 | "is"
 | "jr"
 | "lm"
 | "ez"
 | "dn"
 | "os"
 | "jl"
 | "am"
 | "ob"
 | "jn"
 | "mq"
 | "na"
 | "hc"
 | "sf"
 | "ag"
 | "zc"
 | "ml"
 | "mt"
 | "mc"
 | "lc"
 | "jo"
 | "at"
 | "rm"
 | "1co"
 | "2co"
 | "gl"
 | "ef"
 | "fp"
 | "cl"
 | "1ts"
 | "2ts"
 | "1tm"
 | "2tm"
 | "tt"
 | "fm"
 | "hb"
 | "tg"
 | "1pe"
 | "2pe"
 | "1jo"
 | "2jo"
 | "3jo"
 | "jd"
 | "ap"

export type VerseReference = {
  book: BookAbbr;
  chapter: number;
  verse: number;
}

export type BibleBook = {
  abbrev: BookAbbr;
  chapters: string[][];
};

export type Bible = Array<BibleBook>